requirements.txt
aenum==3.1.15
aiohttp==3.9.5
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
billiard==4.2.0
celery==5.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.2.1
cv==1.0.0
cycler==0.12.1
Deprecated==1.2.14
distro==1.9.0
Django==5.0.6
et-xmlfile==1.1.0
exceptiongroup==1.2.1
filelock==3.14.0
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.5.0
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
intel-openmp==2021.4.0
Jinja2==3.1.4
kiwisolver==1.4.5
kombu==5.3.7
line-bot-sdk==3.11.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mkl==2021.4.0
mpmath==1.3.0
multidict==6.0.5
networkx==3.3
numpy==1.26.4
openai==1.30.1
opencv-python==4.9.0.80
openpyxl==3.1.2
packaging==24.0
pandas==2.2.2
pillow==10.3.0
prompt-toolkit==3.0.43
psutil==5.9.8
py-cpuinfo==9.0.0
pydantic==2.7.1
pydantic_core==2.18.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
scipy==1.13.0
seaborn==0.13.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sqlparse==0.5.0
sympy==1.12
tbb==2021.12.0
thop==0.1.1.post2209072238
torch==2.3.0
torchvision==0.18.0
tqdm==4.66.4
typing_extensions==4.11.0
tzdata==2024.1
ultralytics==8.2.16
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
wrapt==1.16.0
yarl==1.9.4
